+++ src/Internal/RawOutputRecord.cs
@@ -1,13 +1,15 @@
 namespace Sashimi.Internal;
 
-internal abstract record RawOutputRecord(RedirectTo To);
+internal abstract record RawOutputRecord(RedirectTo To, OutputFrom From);
 
-internal record ChunkOutput(byte[] Value, RedirectTo To) : RawOutputRecord(To)
+internal record ChunkOutput(byte[] Value, RedirectTo To, OutputFrom From)
+    : RawOutputRecord(To, From)
 {
     public override string ToString() => string.Join(':', Value.Select(b => b.ToString("X2")));
 }
 
-internal record StringOutput(string Value, RedirectTo To) : RawOutputRecord(To)
+internal record StringOutput(string Value, RedirectTo To, OutputFrom From)
+    : RawOutputRecord(To, From)
 {
     public override string ToString() => Value;
 }