+++ tests/Invoke-RawCommand.tests.ps1
@@ -109,6 +109,43 @@ Describe 'Invoke-RawCommand' {
         }
     }
 
+    Context 'ErrorHandling' {
+        It 'Broken Pipe' {
+            function DelayedOutput { process { Start-Sleep -Milliseconds 100; $input } }
+            $result = 'a' | DelayedOutput | Invoke-RawCommand -ErrorVariable errors -ErrorAction SilentlyContinue cat --invalidParameter
+            $result | Should -BeNullOrEmpty
+            $errors.Count | Should -BeGreaterThan 0
+        }
+    }
+
+    Context 'ThrowOnError' {
+        BeforeAll {
+            ($exe, $arguments) = if ($IsWindows) {
+                'cmd', @('/c', 'exit 1')
+            } else {
+                'sh', @('-c', 'exit 1')
+            }
+        }
+        It 'Not specified `-ThrowOnError` parameter' {
+            { Invoke-RawCommand $exe @arguments } | Should -Not -Throw
+        }
+        It 'Specified `-ThrowOnError` parameter' {
+            try {
+                Invoke-RawCommand -ThrowOnError $exe @arguments
+            } catch {
+                $err = $_
+            }
+
+            $err | Should -Not -BeNullOrEmpty
+
+            $err.FullyQualifiedErrorId | Should -BeLike 'ExternalCommandNonZeroExit,*'
+            $err.CategoryInfo.Category | Should -Be 'InvalidResult'
+
+            $err.Exception | Should -BeOfType [Sashimi.ExternalCommandNonZeroExitException]
+            $err.Exception.ExitCode | Should -BeGreaterThan 0
+        }
+    }
+
     Context 'Cancel' {
         BeforeAll {
             $env:SashimiModulePath = Resolve-Path -RelativeBasePath $PSScriptRoot -Path ..